requirements.txt
streamlit
customtkinter
google-generativeai